fix(styles): moves CSS out of published project
#41

- Move all CSS imports into `styles.ts`
- Group all Storybook content in `storybook` folder
- Ensure no CSS imports outside of Storybook
