+++ src/stories/flows/DesignMode.stories.tsx
@@ -1,10 +1,9 @@
 import { Meta, Story } from '@storybook/react';
-import '@trussworks/react-uswds/lib/index.css';
-import '@trussworks/react-uswds/lib/uswds.css';
-import { Questionable } from '../components/Questionable';
+import { Questionable } from '../../components/Questionable';
 import { simpleFlow } from '../flows/simple/simple.flow';
-import { QuestionableConfig } from '../composable/Config';
-import { IQuestionable } from '../state/GlobalState';
+import { MODE } from '../../lib';
+import { QuestionableConfig } from '../../composable/Config';
+import { IQuestionable } from '../../state/GlobalState';
 import '../styles';
 
 export default {
@@ -12,15 +11,16 @@ export default {
     config: { control: { type: 'object' } },
   },
   component: Questionable,
-  title: 'Questionable/Simple Flow',
+  title: 'Questionable/Design Mode',
 } as Meta;
 
 const Template: Story<IQuestionable> = (args) => <Questionable {...args} />;
 
-export const Simple = Template.bind({});
-Simple.args = {
+export const Design = Template.bind({});
+Design.args = {
   config: new QuestionableConfig({
     dev: false,
+    mode: MODE.EDIT,
     steps: {
       showStepId: false,
     }
