+++ src/stories/flows/complex/ComplexFlow.stories.tsx
@@ -1,11 +1,9 @@
 import { Meta, Story } from '@storybook/react';
-import '@trussworks/react-uswds/lib/index.css';
-import '@trussworks/react-uswds/lib/uswds.css';
-import { Questionable } from '../components/Questionable';
-import { complexFlow } from '../flows/complex/complex.flow';
-import { QuestionableConfig } from '../composable/Config';
-import { IQuestionable } from '../state/GlobalState';
-import '../styles';
+import { Questionable } from '../../../components/Questionable';
+import { complexFlow } from './complex.flow';
+import { QuestionableConfig } from '../../../composable/Config';
+import { IQuestionable } from '../../../state/GlobalState';
+import '../../styles';
 
 export default {
   argTypes: {
