--- src/styles/index.ts
@@ -1,1 +1,0 @@
-export * from './index.css';
