+++ tsconfig.build.json
@@ -2,35 +2,48 @@
   "compilerOptions": {
     "allowJs": true,
     "allowSyntheticDefaultImports": true,
+    "alwaysStrict": true,
+    "declaration": true,
+    "declarationMap": true,
     "downlevelIteration": true,
     "esModuleInterop": true,
     "experimentalDecorators": true,
     "forceConsistentCasingInFileNames": true,
     "isolatedModules": true,
     "jsx": "react-jsx",
-    "lib": ["dom", "dom.iterable", "esnext"],
-    "module": "esnext",
+    "lib": [
+      "dom",
+      "ES2015"
+    ],
+    "module": "CommonJS",
     "moduleResolution": "node",
     "noEmit": true,
     "noFallthroughCasesInSwitch": true,
+    "noImplicitReturns": true,
+    "noImplicitThis": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
     "outDir": "dist",
     "resolveJsonModule": true,
     "skipLibCheck": true,
-    "strict": true,
-    "target": "ESNext",
-    "declaration": true,
-    "declarationMap": true,
     "sourceMap": true,
-    "strictNullChecks": true,
+    "strict": true,
     "strictFunctionTypes": true,
+    "strictNullChecks": true,
     "strictPropertyInitialization": true,
-    "noImplicitThis": true,
-    "alwaysStrict": true,
-    "noUnusedLocals": true,
-    "noUnusedParameters": true,
-    "noImplicitReturns": true
+    "target": "ES2015"
   },
-  "include": ["src"],
-  "exclude": ["node_modules"],
-  "files": ["src/index.ts"]
+  "exclude": [
+    "node_modules",
+    "src/stories",
+    "**/*.stories.ts",
+    "**/*.css",
+    "**/*.test.ts"
+  ],
+  "files": [
+    "src/index.ts"
+  ],
+  "include": [
+    "src"
+  ]
 }
